# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudAI Fusion API
  description: |
    CloudAI Fusion - Cloud-Native AI Unified Management Platform API.
    Provides endpoints for multi-cloud cluster management, AI workload scheduling,
    security policy enforcement, monitoring & alerting, and cost optimization.
  version: 0.1.0-mvp
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: CloudAI Fusion Team
    url: https://github.com/cloudai-fusion/cloudai-fusion
servers:
  - url: http://localhost:8080
    description: Go API Server (clusters, auth, security, monitoring)
  - url: http://localhost:8090
    description: Python AI Engine (agents, LLM, chat, operations)
  - url: https://api.cloudai-fusion.io
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: User login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh JWT token
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/v1/evidence/records:
    get:
      summary: List signed receipts (newest first)
      tags: [evidence]
      parameters:
        - {name: action, in: query, schema: {type: string}}
        - {name: subject, in: query, schema: {type: string}}
        - {name: actor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer}}
      responses:
        "200":
          description: Receipts
  /api/v1/evidence/records/{id}:
    get:
      summary: Get a single receipt
      tags: [evidence]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200": {description: Receipt}
        "404": {description: Not found}
  /api/v1/evidence/records/{id}/proof:
    get:
      summary: Merkle inclusion proof for a receipt
      tags: [evidence]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200": {description: Inclusion proof}
...
  /api/v1/redteam/engagements/{id}/evidence:
    get:
      summary: Signed receipts for an engagement
      tags: [redteam]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200": {description: Receipts}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          type: object

    # ==================== Edge AI Schemas ====================
...